requirements.txt
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.3.2
google-ads==21.2.0
google-api-core==2.11.0
google-auth==2.20.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.1
grpcio==1.54.2
grpcio-status==1.54.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
oauthlib==3.2.2
proto-plus==1.22.2
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
urllib3==1.26.16
Werkzeug==2.3.6
